Moved and changed privacy on map
