+++ core/src/com/g16/feyrune/model/Parser/map/Tile.java
@@ -1,4 +1,4 @@
-package com.g16.feyrune.map;
+package com.g16.feyrune.model.Parser.map;
 
 public class Tile {
     private final boolean collision;
