+++ desktop/src/com/g16/feyrune/DesktopLauncher.java
@@ -2,9 +2,6 @@ package com.g16.feyrune;
 
 import com.badlogic.gdx.backends.lwjgl3.Lwjgl3Application;
 import com.badlogic.gdx.backends.lwjgl3.Lwjgl3ApplicationConfiguration;
-import com.g16.feyrune.Feyrune;
-import com.g16.feyrune.map.Map;
-import com.g16.feyrune.model.Parser.MapParser;
 
 // Please note that on macOS your application needs to be started with the -XstartOnFirstThread JVM argument
 public class DesktopLauncher {
